services/instagram-scraper/pyproject.toml
[project]
name = "instagram-scraper"
version = "0.1.0"
description = "Instagram scraper service for recipe data extraction"
requires-python = ">=3.12"
dependencies = [
    "celery[redis]>=5.3.4",
    "instaloader>=4.10.3",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-celery>=0.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
exclude = ["tests/", ".venv/", "__pycache__/"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.black]
line-length = 88
target-version = ["py312"]

[dependency-groups]
dev = [
    "black>=25.11.0",
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.5",
]
